Cookie service was increased.
Users needed to remain logged in while they are exploring. When the user is logged in, they do not have to sign in to continue using their page. Can use /credentials/authorize to ensure that the user is still logged in. There is also the implementation of logout that will delete the JWT token from local storage and login is locked. If the code does not work, run an NPM I in the main and frontend directories.