Comments; changes to getUserFunds() in GO. Modularized claim extraction
Added comments to almost all the functions to make them easier to understand. JWTAuthProtection to JWTPathProtecion to reflect the usage. Users must be logged in to continue to the http request, otherwise returns 401.  GetUserFunds in GO was changed to use httponly cookie to get username and is now Path protected. This does not affect the the functionality. It still returns funds. Lastly, new extractclaims that can be used to extract cookie claims from any other part in GO. Does not need to go through user log in verification. Use under path protected endpoints.